api/requirements.txt
annotated-types==0.7.0
alembic==1.13.2
anyio==4.4.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.5
fastjsonschema==2.19.1
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.7
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.5.0
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
passlib[bcrypt]==1.7.4
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rapidfuzz==3.9.1
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.8.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
tomlkit==0.12.5
tqdm==4.66.5
trove-classifiers==2024.5.22
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.2
watchfiles==0.24.0
websockets==13.0.1
xattr==1.1.0
any-llm-sdk[openai,anthropic]

